requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
langgraph>=0.1.4
langchain-google-genai>=1.0.6
google-genai>=0.1.0
google-generativeai>=0.7.0 
# Keeping google-generativeai for langchain compatibility if needed, but adding google-genai
pydantic>=2.7.4
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
tenacity>=8.4.1
email-validator>=2.2.0
httpx>=0.27.0
regex>=2024.5.15
typing-extensions>=4.12.2
